+++ templates/react-h5/postcss.config.js
@@ -1,0 +1,22 @@
+module.exports = ctx => ({
+  parser: ctx.parser ? 'sugarss' : false,
+  map: ctx.env === 'development' ? ctx.map : false,
+  plugins: {
+    'postcss-nested': {},
+    'postcss-import': {},
+    'postcss-url': {},
+    'postcss-aspect-ratio-mini': {},
+    'postcss-write-svg': { utf8: false },
+    'postcss-preset-env': {},
+    'postcss-px-to-viewport': {
+      viewportWidth: 750, // (Number) The width of the viewport.
+      viewportHeight: 1334, // (Number) The height of the viewport.
+      unitPrecision: 3, // (Number) The decimal numbers to allow the REM units to grow to.
+      viewportUnit: 'vw', // (String) Expected units.
+      selectorBlackList: ['.ignore', '.hairlines'], // (Array) The selectors to ignore and leave as px.
+      minPixelValue: 1, // (Number) Set the minimum pixel value to replace.
+      mediaQuery: false, // (Boolean) Allow px to be converted in media queries.
+    },
+    // 'postcss-viewport-units': {}, // 自动添加content for viewport-units-buggyfill. 
+  },
+});